requirements.txt
pillow
mysql-connector-python
python-dotenv